+++ climpred/metrics.py
@@ -8,11 +8,16 @@ from xskillscore import (
     crps_ensemble,
     crps_gaussian,
     crps_quadrature,
+    mad,
     mae,
+    mape,
     mse,
     pearson_r,
     pearson_r_p_value,
     rmse,
+    smape,
+    spearman_r,
+    spearman_r_p_value,
     threshold_brier_score,
 )
 
@@ -46,15 +51,22 @@ def _get_norm_factor(comparison):
     return fac
 
 
-def _pearson_r(forecast, reference, dim='svd', **metric_kwargs):
+def _pearson_r(forecast, reference, dim=None, **metric_kwargs):
     """
-    Calculate the Anomaly Correlation Coefficient (ACC).
+    Calculate Person's Anomaly Correlation Coefficient (ACC).
 
     .. math::
         ACC = \\frac{cov(f, o)}{\\sigma_{f}\\cdot\\sigma_{o}}
 
     .. note::
         Use metric ``pearson_r_p_value`` to get the corresponding pvalue.
+
+    Args:
+        forecast (xarray object): forecast
+        reference (xarray object): reference
+        dim (str): dimension(s) to perform metric over.
+                   Automatically set by compute_.
+        metric_kwargs: (optional) weights, skipna, see xskillscore.pearson_r
 
     Range:
         * perfect: 1
@@ -64,28 +76,116 @@ def _pearson_r(forecast, reference, dim='svd', **metric_kwargs):
         * xskillscore.pearson_r
         * xskillscore.pearson_r_p_value
     """
-    return pearson_r(forecast, reference, dim=dim)
+    weights = metric_kwargs.get('weights', None)
+    skipna = metric_kwargs.get('skipna', False)
+    return pearson_r(forecast, reference, dim=dim, weights=weights, skipna=skipna)
 
 
-def _pearson_r_p_value(forecast, reference, dim='svd', **metric_kwargs):
+# TODO: all metrics: docs
+def _pearson_r_p_value(forecast, reference, dim=None, **metric_kwargs):
     """
-    Calculate the probability associated with the ACC not being random.
+    Calculate the probability associated with Person's Anomaly Correlation
+    Coefficient not being random.
+
+    Args:
+        forecast (xarray object): forecast
+        reference (xarray object): reference
+        dim (str): dimension(s) to perform metric over.
+                   Automatically set by compute_.
+        metric_kwargs: (optional) weights, skipna, see xskillscore.pearson_r_p_value
+
+    Range:
+        * perfect: 0
+        * max: 1
+
+    See also:
+        * xskillscore.pearson_r_p_value
     """
+    weights = metric_kwargs.get('weights', None)
+    skipna = metric_kwargs.get('skipna', False)
     # p-value returns a runtime error when working with NaNs, such as on a climate
     # model grid. We can avoid this annoying output by specifically suppressing
     # warning here.
     with warnings.catch_warnings():
         warnings.simplefilter('ignore')
-        pval = pearson_r_p_value(forecast, reference, dim=dim)
-    return pval
+        return pearson_r_p_value(
+            forecast, reference, dim=dim, weights=weights, skipna=skipna
+        )
 
 
-def _mse(forecast, reference, dim='svd', **metric_kwargs):
+def _spearman_r(forecast, reference, dim=None, **metric_kwargs):
+    """
+    Calculate Spearman's Anomaly Correlation Coefficient (SACC).
+
+    .. math::
+        SACC = ACC(ranked(f),ranked(o))
+
+    .. note::
+        Use metric ``spearman_r_p_value`` to get the corresponding pvalue.
+
+    Args:
+        forecast (xarray object): forecast
+        reference (xarray object): reference
+        dim (str): dimension(s) to perform metric over.
+                   Automatically set by compute_.
+        metric_kwargs: (optional) weights, skipna, see xskillscore.spearman_r
+
+    Range:
+        * perfect: 1
+        * min: -1
+
+    See also:
+        * xskillscore.spearman_r
+        * xskillscore.spearman_r_p_value
+    """
+    weights = metric_kwargs.get('weights', None)
+    skipna = metric_kwargs.get('skipna', False)
+    return spearman_r(forecast, reference, dim=dim, weights=weights, skipna=skipna)
+
+
+def _spearman_r_p_value(forecast, reference, dim=None, **metric_kwargs):
+    """
+    Calculate the probability associated with the ACC not being random.
+
+    Args:
+        forecast (xarray object): forecast
+        reference (xarray object): reference
+        dim (str): dimension(s) to perform metric over.
+                   Automatically set by compute_.
+        metric_kwargs: (optional) weights, skipna, see xskillscore.spearman_r_p_value
+
+    Range:
+        * perfect: 0
+        * max: 1
+
+    See also:
+        * xskillscore.spearman_r_p_value
+    """
+    weights = metric_kwargs.get('weights', None)
+    skipna = metric_kwargs.get('skipna', False)
+    # p-value returns a runtime error when working with NaNs, such as on a climate
+    # model grid. We can avoid this annoying output by specifically suppressing
+    # warning here.
+    with warnings.catch_warnings():
+        warnings.simplefilter('ignore')
+        return spearman_r_p_value(
+            forecast, reference, dim=dim, weights=weights, skipna=skipna
+        )
+
+
+def _mse(forecast, reference, dim=None, **metric_kwargs):
     """
     Calculate the Mean Sqaure Error (MSE).
 
     .. math::
         MSE = \\overline{(f - o)^{2}}
+
+    Args:
+        forecast (xarray object): forecast
+        reference (xarray object): reference
+        dim (str): dimension(s) to perform metric over.
+                   Automatically set by compute_.
+        metric_kwargs: (optional) weights, skipna, see xskillscore.mse
 
     Range:
         * perfect: 0
@@ -95,15 +195,24 @@ def _mse(forecast, reference, dim='svd', **metric_kwargs):
     See also:
         * xskillscore.mse
     """
-    return mse(forecast, reference, dim=dim)
+    weights = metric_kwargs.get('weights', None)
+    skipna = metric_kwargs.get('skipna', False)
+    return mse(forecast, reference, dim=dim, weights=weights, skipna=skipna)
 
 
-def _rmse(forecast, reference, dim='svd', **metric_kwargs):
+def _rmse(forecast, reference, dim=None, **metric_kwargs):
     """
     Calculate the Root Mean Sqaure Error (RMSE).
 
     .. math::
         RMSE = \\sqrt{\\overline{(f - o)^{2}}}
+
+    Args:
+        forecast (xarray object): forecast
+        reference (xarray object): reference
+        dim (str): dimension(s) to perform metric over.
+                   Automatically set by compute_.
+        metric_kwargs: (optional) weights, skipna, see xskillscore.rmse
 
     Range:
         * perfect: 0
@@ -113,15 +222,24 @@ def _rmse(forecast, reference, dim='svd', **metric_kwargs):
     See also:
         * xskillscore.rmse
     """
-    return rmse(forecast, reference, dim=dim)
+    weights = metric_kwargs.get('weights', None)
+    skipna = metric_kwargs.get('skipna', False)
+    return rmse(forecast, reference, dim=dim, weights=weights, skipna=skipna)
 
 
-def _mae(forecast, reference, dim='svd', **metric_kwargs):
+def _mae(forecast, reference, dim=None, **metric_kwargs):
     """
     Calculate the Mean Absolute Error (MAE).
 
     .. math::
-        MSE = \\overline{(f - o)^{2}}
+        MAE = \\overline{|f - o|}
+
+    Args:
+        forecast (xarray object): forecast
+        reference (xarray object): reference
+        dim (str): dimension(s) to perform metric over.
+                   Automatically set by compute_.
+        metric_kwargs: (optional) weights, skipna, see xskillscore.mae
 
     Range:
         * perfect: 0
@@ -131,7 +249,89 @@ def _mae(forecast, reference, dim='svd', **metric_kwargs):
     See also:
         * xskillscore.mae
     """
-    return mae(forecast, reference, dim=dim)
+    weights = metric_kwargs.get('weights', None)
+    skipna = metric_kwargs.get('skipna', False)
+    return mae(forecast, reference, dim=dim, weights=weights, skipna=skipna)
+
+
+def _mad(forecast, reference, dim=None, **metric_kwargs):
+    """
+    Calculate the Median Absolute Deviation (MAD).
+
+    .. math::
+        MAD = median(|f - o|)
+
+    Args:
+        forecast (xarray object): forecast
+        reference (xarray object): reference
+        dim (str): dimension(s) to perform metric over.
+                   Automatically set by compute_.
+        metric_kwargs: (optional) weights, skipna, see xskillscore.mad
+
+    Range:
+        * perfect: 0
+        * min: 0
+        * max: ∞
+
+    See also:
+        * xskillscore.mad
+    """
+    skipna = metric_kwargs.get('skipna', False)
+    return mad(forecast, reference, dim=dim, skipna=skipna)
+
+
+def _mape(forecast, reference, dim=None, **metric_kwargs):
+    """
+    Calculate the Mean Absolute Percentage Error (MAPE).
+
+    .. math::
+        MAPE = MAPE = 1/n \sum \frac{|f-o|}{|o|}
+
+    Args:
+        forecast (xarray object): forecast
+        reference (xarray object): reference
+        dim (str): dimension(s) to perform metric over.
+                   Automatically set by compute_.
+        metric_kwargs: (optional) weights, skipna, see xskillscore.mape
+
+    Range:
+        * perfect: 0
+        * min: 0
+        * max: ∞
+
+    See also:
+        * xskillscore.mape
+    """
+    weights = metric_kwargs.get('weights', None)
+    skipna = metric_kwargs.get('skipna', False)
+    return mape(forecast, reference, dim=dim, weights=weights, skipna=skipna)
+
+
+def _smape(forecast, reference, dim=None, **metric_kwargs):
+    """
+    Calculate the symmetric Mean Absolute Percentage Error (sMAPE).
+
+    .. math::
+        sMAPE = 1/n \sum \frac{|f-o|}{|f|+|o|}
+
+    Args:
+        forecast (xarray object): forecast
+        reference (xarray object): reference
+        dim (str): dimension(s) to perform metric over.
+                   Automatically set by compute_.
+        metric_kwargs: (optional) weights, skipna, see xskillscore.smape
+
+    Range:
+        * perfect: 0
+        * min: 0
+        * max: ∞
+
+    See also:
+        * xskillscore.smape
+    """
+    weights = metric_kwargs.get('weights', None)
+    skipna = metric_kwargs.get('skipna', False)
+    return smape(forecast, reference, dim=dim, weights=weights, skipna=skipna)
 
 
 def _brier_score(forecast, reference, **metric_kwargs):
@@ -141,8 +341,8 @@ def _brier_score(forecast, reference, **metric_kwargs):
         BS(f, o) = (f - o)^2
 
     Args:
-        * forecast (xr.object)
-        * reference (xr.object)
+        * forecast (xr.object): forecast with `member` dim
+        * reference (xr.object): references without `member` dim
         * func (function): function to be applied to reference and forecasts
                            and then mean('member') to get forecasts and
                            reference in interval [0,1].
@@ -186,8 +386,8 @@ def _threshold_brier_score(forecast, reference, **metric_kwargs):
         * max: 1
 
     Args:
-        * forecast (xr.object)
-        * reference (xr.object)
+        * forecast (xr.object): forecast with `member` dim
+        * reference (xr.object): references without `member` dim
         * threshold (int, float, xr.object): Threshold to check exceedance,
             see properscoring.threshold_brier_score
             (required to be added via **metric_kwargs)
@@ -217,6 +417,11 @@ def _crps(forecast, reference, **metric_kwargs):
     """
     Continuous Ranked Probability Score (CRPS) is the probabilistic MSE.
 
+    Args:
+        * forecast (xr.object): forecast with `member` dim
+        * reference (xr.object): references without `member` dim
+        * metric_kwargs (optional dict): weights, see properscoring.crps_ensemble
+
     Range:
         * perfect: 0
         * min: 0
@@ -232,11 +437,17 @@ def _crps(forecast, reference, **metric_kwargs):
         * xskillscore.crps_ensemble
     """
     # switch positions because xskillscore.crps_ensemble(obs, forecasts)
-    return crps_ensemble(reference, forecast)
+    weights = metric_kwargs.get('weights', None)
+    return crps_ensemble(reference, forecast, weights=weights)
 
 
 def _crps_gaussian(forecast, mu, sig, **metric_kwargs):
     """CRPS assuming a gaussian distribution. Helper function for CRPSS.
+
+    Args:
+        * forecast (xr.object): forecast with `member` dim
+        * mu (xr.object): mean reference
+        * sig (xr.object): standard deviation reference
 
     See also:
         * properscoring.crps_gaussian
@@ -249,6 +460,10 @@ def _crps_quadrature(
     forecast, cdf_or_dist, xmin=None, xmax=None, tol=1e-6, **metric_kwargs
 ):
     """CRPS assuming distribution cdf_or_dist. Helper function for CRPSS.
+
+    Args:
+        * forecast (xr.object): forecast with `member` dim
+        * see properscoring.crps_quadrature
 
     See also:
         * properscoring.crps_quadrature
@@ -269,8 +484,8 @@ def _crpss(forecast, reference, **metric_kwargs):
         CRPSS = 1 - \\frac{CRPS_{init}}{CRPS_{clim}}
 
     Args:
-        * forecast (xr.object):
-        * reference (xr.object):
+        * forecast (xr.object): forecast with `member` dim
+        * reference (xr.object): references without `member` dim
         * gaussian (bool): Assuming gaussian distribution for baseline skill.
                            Default: True (optional)
         * cdf_or_dist (scipy.stats): distribution to assume if not gaussian.
@@ -332,7 +547,7 @@ def _crpss(forecast, reference, **metric_kwargs):
         else:
             tol = 1e-6
         ref_skill = _crps_quadrature(forecast, cdf_or_dist, xmin, xmax, tol)
-    forecast_skill = _crps(forecast, reference)
+    forecast_skill = _crps(forecast, reference, **metric_kwargs)
     skill_score = 1 - forecast_skill / ref_skill.mean('member')
     return skill_score
 
@@ -341,6 +556,11 @@ def _crpss_es(forecast, reference, **metric_kwargs):
     """CRPSS Ensemble Spread.
 
     .. math:: CRPSS = 1 - \\frac{CRPS(\\sigma^2_f)}{CRPS(\\sigma^2_o}))
+
+    Args:
+        * forecast (xr.object): forecast with `member` dim
+        * reference (xr.object): references without `member` dim
+        * metric_kwargs (optional): weights, skipna used for mse
 
     References:
         * Kadow, Christopher, Sebastian Illing, Oliver Kunst, Henning W. Rust,
@@ -366,8 +586,11 @@ def _crpss_es(forecast, reference, **metric_kwargs):
 
     mu = reference.mean(rdim)
     forecast, ref2 = xr.broadcast(forecast, reference)
-    sig_r = _mse(forecast, ref2, dim='member').mean(dim2)
-    sig_h = _mse(forecast.mean(dim2), ref2.mean(dim2), 'member')
+    mse_kwargs = metric_kwargs.copy()
+    if 'dim' in mse_kwargs:
+        del mse_kwargs['dim']
+    sig_r = _mse(forecast, ref2, dim='member', **mse_kwargs).mean(dim2)
+    sig_h = _mse(forecast.mean(dim2), ref2.mean(dim2), dim='member', **mse_kwargs)
     crps_h = _crps_gaussian(forecast, mu, sig_h)
     if 'member' in crps_h.dims:
         crps_h = crps_h.mean('member')
@@ -377,11 +600,17 @@ def _crpss_es(forecast, reference, **metric_kwargs):
     return 1 - crps_h / crps_r
 
 
-def _bias(forecast, reference, dim='svd', **metric_kwargs):
+def _bias(forecast, reference, dim=None, **metric_kwargs):
     """Calculate unconditional bias.
 
     .. math::
         bias = f - o
+
+    Args:
+        forecast (xarray object): forecast
+        reference (xarray object): reference
+        dim (str): dimension(s) to perform metric over.
+                   Automatically set by compute_.
 
     Range:
         * pos: positive bias
@@ -396,12 +625,19 @@ def _bias(forecast, reference, dim='svd', **metric_kwargs):
     return bias
 
 
-def _msss_murphy(forecast, reference, dim='svd', **metric_kwargs):
+def _msss_murphy(forecast, reference, dim=None, **metric_kwargs):
     """Calculate Murphy's Mean Square Skill Score (MSSS).
 
     .. math::
         MSSS_{Murphy} = r_{fo}^2 - [\\text{conditional bias}]^2 -\
          [\\frac{\\text{(unconditional) bias}}{\\sigma_o}]^2
+
+    Args:
+        forecast (xarray object): forecast
+        reference (xarray object): reference
+        dim (str): dimension(s) to perform metric over.
+                   Automatically set by compute_.
+        metric_kwargs: (optional) weights, skipna
 
     References:
         * https://www-miklip.dkrz.de/about/murcss/
@@ -410,30 +646,46 @@ def _msss_murphy(forecast, reference, dim='svd', **metric_kwargs):
           Review 116, no. 12 (December 1, 1988): 2417–24.
           https://doi.org/10/fc7mxd.
     """
-    acc = _pearson_r(forecast, reference, dim=dim)
-    conditional_bias = _conditional_bias(forecast, reference, dim=dim)
-    uncond_bias = _bias(forecast, reference, dim=dim) / reference.std(dim)
+    acc = _pearson_r(forecast, reference, dim=dim, **metric_kwargs)
+    conditional_bias = _conditional_bias(forecast, reference, dim=dim, **metric_kwargs)
+    uncond_bias = _bias(forecast, reference, dim=dim, **metric_kwargs) / reference.std(
+        dim
+    )
     skill = acc ** 2 - conditional_bias ** 2 - uncond_bias ** 2
     return skill
 
 
-def _conditional_bias(forecast, reference, dim='svd', **metric_kwargs):
+def _conditional_bias(forecast, reference, dim=None, **metric_kwargs):
     """Calculate the conditional bias between forecast and reference.
 
     .. math:: \\text{conditional bias} = r_{fo} - \\frac{\\sigma_f}{\\sigma_o}
 
+    Args:
+        forecast (xarray object): forecast
+        reference (xarray object): reference
+        dim (str): dimension(s) to perform metric over.
+                   Automatically set by compute_.
+
     References:
         * https://www-miklip.dkrz.de/about/murcss/
     """
-    acc = _pearson_r(forecast, reference, dim=dim)
-    conditional_bias = acc - _std_ratio(forecast, reference, dim=dim) ** -1
+    acc = _pearson_r(forecast, reference, dim=dim, **metric_kwargs)
+    conditional_bias = (
+        acc - _std_ratio(forecast, reference, dim=dim, **metric_kwargs) ** -1
+    )
     return conditional_bias
 
 
-def _std_ratio(forecast, reference, dim='svd', **metric_kwargs):
+def _std_ratio(forecast, reference, dim=None, **metric_kwargs):
     """Calculate the ratio of standard deviations of reference over forecast.
 
     .. math:: \\text{std ratio} = \\frac{\\sigma_o}{\\sigma_f}
+
+    Args:
+        forecast (xarray object): forecast
+        reference (xarray object): reference
+        dim (str): dimension(s) to perform metric over.
+                   Automatically set by compute_.
 
     References:
         * https://www-miklip.dkrz.de/about/murcss/
@@ -442,31 +694,39 @@ def _std_ratio(forecast, reference, dim='svd', **metric_kwargs):
     return ratio
 
 
-def _bias_slope(forecast, reference, dim='svd', **metric_kwargs):
+def _bias_slope(forecast, reference, dim=None, **metric_kwargs):
     """Calculate bias slope between reference and forecast standard deviations.
 
     .. math:: \\text{bias slope}= r_{fo} \\cdot \\text{std ratio}
 
+    Args:
+        forecast (xarray object): forecast
+        reference (xarray object): reference
+        dim (str): dimension(s) to perform metric over.
+                   Automatically set by compute_.
+
     References:
         * https://www-miklip.dkrz.de/about/murcss/
     """
-    std_ratio = _std_ratio(forecast, reference, dim=dim)
-    acc = _pearson_r(forecast, reference, dim=dim)
+    std_ratio = _std_ratio(forecast, reference, dim=dim, **metric_kwargs)
+    acc = _pearson_r(forecast, reference, dim=dim, **metric_kwargs)
     b_s = std_ratio * acc
     return b_s
 
 
-def _ppp(forecast, reference, dim='svd', **metric_kwargs):
+def _ppp(forecast, reference, dim=None, **metric_kwargs):
     """Prognostic Potential Predictability (PPP) metric.
 
     .. math:: PPP = 1 - \\frac{MSE}{ \\sigma^2_{ref} \\cdot fac}
 
     Args:
-        * forecast (xr.object)
-        * reference (xr.object)
-        * dim (str): dimension to apply metric to
-        * comparison (str): name comparison needed for normalization factor
+        forecast (xarray object): forecast
+        reference (xarray object): reference
+        dim (str): dimension(s) to perform metric over.
+                   Automatically set by compute_.
+        comparison (str): name comparison needed for normalization factor
                            (required to be added via **metric_kwargs)
+        metric_kwargs: (optional) weights, skipna, see xskillscore.mse
 
     Range:
         * 1: perfect forecast
@@ -486,7 +746,7 @@ def _ppp(forecast, reference, dim='svd', **metric_kwargs):
         Prediction: Potential versus Operational Seasonal Forecast Skill.
         Climate Dynamics, June 9, 2018. https://doi.org/10/gd7hfq.
     """
-    mse_skill = _mse(forecast, reference, dim=dim)
+    mse_skill = _mse(forecast, reference, dim=dim, **metric_kwargs)
     var = reference.var(dim)
     if 'comparison' in metric_kwargs:
         comparison = metric_kwargs['comparison']
@@ -499,7 +759,7 @@ def _ppp(forecast, reference, dim='svd', **metric_kwargs):
     return ppp_skill
 
 
-def _nrmse(forecast, reference, dim='svd', **metric_kwargs):
+def _nrmse(forecast, reference, dim=None, **metric_kwargs):
     """Normalized Root Mean Square Error (NRMSE) metric.
 
     .. math:: NRMSE = \\frac{RMSE}{\\sigma_{o} \\cdot \\sqrt{fac} }
@@ -529,7 +789,7 @@ def _nrmse(forecast, reference, dim='svd', **metric_kwargs):
         (January 1, 2016): 672–83. https://doi.org/10/gfb3pn.
 
     """
-    rmse_skill = _rmse(forecast, reference, dim=dim)
+    rmse_skill = _rmse(forecast, reference, dim=dim, **metric_kwargs)
     std = reference.std(dim)
     if 'comparison' in metric_kwargs:
         comparison = metric_kwargs['comparison']
@@ -542,7 +802,7 @@ def _nrmse(forecast, reference, dim='svd', **metric_kwargs):
     return nrmse_skill
 
 
-def _nmse(forecast, reference, dim='svd', **metric_kwargs):
+def _nmse(forecast, reference, dim=None, **metric_kwargs):
     """
     Calculate Normalized MSE (NMSE) = Normalized Ensemble Variance (NEV).
 
@@ -565,7 +825,7 @@ def _nmse(forecast, reference, dim='svd', **metric_kwargs):
           North Atlantic Multidecadal Variability.” Climate Dynamics 13,
           no. 7–8 (August 1, 1997): 459–87. https://doi.org/10/ch4kc4.
     """
-    mse_skill = _mse(forecast, reference, dim=dim)
+    mse_skill = _mse(forecast, reference, dim=dim, **metric_kwargs)
     var = reference.var(dim)
     if 'comparison' in metric_kwargs:
         comparison = metric_kwargs['comparison']
@@ -578,7 +838,7 @@ def _nmse(forecast, reference, dim='svd', **metric_kwargs):
     return nmse_skill
 
 
-def _nmae(forecast, reference, dim='svd', **metric_kwargs):
+def _nmae(forecast, reference, dim=None, **metric_kwargs):
     """
     Normalized Ensemble Mean Absolute Error metric.
 
@@ -602,7 +862,7 @@ def _nmae(forecast, reference, dim='svd', **metric_kwargs):
           7–8 (August 1, 1997): 459–87. https://doi.org/10/ch4kc4.
 
     """
-    mae_skill = _mae(forecast, reference, dim=dim)
+    mae_skill = _mae(forecast, reference, dim=dim, **metric_kwargs)
     std = reference.std(dim).mean()
     if 'comparison' in metric_kwargs:
         comparison = metric_kwargs['comparison']
@@ -615,11 +875,18 @@ def _nmae(forecast, reference, dim='svd', **metric_kwargs):
     return nmae_skill
 
 
-def _uacc(forecast, reference, dim='svd', **metric_kwargs):
+def _uacc(forecast, reference, dim=None, **metric_kwargs):
     """
     Calculate Bushuk's unbiased ACC (uACC).
 
     .. math:: uACC = \\sqrt{PPP} = \\sqrt{MSSS}
+
+    Args:
+        * forecast (xr.object)
+        * reference (xr.object)
+        * dim (str): dimension to apply metric to
+        * metric_kwargs: (optional) weights, skipna, see xskillscore.mse
+
 
     Range:
         * 1: perfect
@@ -632,4 +899,7 @@ def _uacc(forecast, reference, dim='svd', **metric_kwargs):
           Forecast Skill. Climate Dynamics, June 9, 2018.
           https://doi.org/10/gd7hfq.
     """
-    return _ppp(forecast, reference, dim=dim, **metric_kwargs) ** 0.5
+    ppp_res = _ppp(forecast, reference, dim=dim, **metric_kwargs)
+    # ensure no sqrt of neg values
+    uacc_res = (ppp_res.where(ppp_res > 0)) ** 0.5
+    return uacc_res