+++ CHANGELOG.rst
@@ -5,8 +5,15 @@ Changelog History
 climpred v1.1.1 (2019-##-##)
 ============================
 
+Depreciated
+-----------
+- Abbreviation `pval` depreciated. Use `p_pval` for `pearson_r_p_value` instead. (:pr:`264`) `Aaron Spring`_.
+
+
 Features
 --------
+- `metric_kwargs` are passed to `xs.metric`. (:pr:`264`) `Aaron Spring`_.
+- New deterministic metrics: `spearman_r`, `spearman_r_p_value`, `mad`, `mape`, `smape`. (:pr:`264`) `Aaron Spring`_.
 - Rewrote `varweighted_mean_period` based on `xrft`. Changed `time_dim` to `dim` in `stats.varweighted_mean_period`. Keeps coords. (:pr:`258`) `Aaron Spring`_.
 - Add `dim='time'` in `stats.dpp` (:pr:`258`) `Aaron Spring`_.
 - Apply arbitrary ``xarray`` methods to ``HindcastEnsemble`` and ``PerfectModelEnsemble`` (:pr:`243`) `Riley X. Brady`_.
@@ -15,9 +22,11 @@ Features
 Bug Fixes
 ---------
 - Add `skill` from `compute_hindcast` keeps coords of `hindcast` (:pr:`258`) `Aaron Spring`_.
+- metric `uacc` does not crash when `ppp` negative. (:pr:`264`) `Aaron Spring`_.
 
 Internals/Minor Fixes
 ---------------------
+- Comparisons `m2m`, `m2e` rewritten to not stack dims into supervector dim because this happens now in `xskillscore`. (:pr:`264`) `Aaron Spring`_.
 - Add ``tqdm`` progress bar to ``bootstrap_compute`` (:pr:`244`) `Aaron Spring`_.
 - Remove inplace behavior for ``HindcastEnsemble`` and ``PerfectModelEnsemble`` (:pr:`243`) `Riley X. Brady`_.
 - Added tests for chunking with `dask` (:pr:`258`) `Aaron Spring`_.